requirements.txt
aiohttp==3.6.2
appdirs==1.4.0
async-timeout==3.0.1
attrs==19.3.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
Flask==0.12.3
idna==2.9
idna-ssl==1.1.0
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
multidict==4.7.5
packaging==16.8
pyee==5.0.0
pyparsing==2.1.10
requests==2.23.0
six==1.10.0
slackclient==2.5.0
slackeventsapi==2.1.0
typing-extensions==3.7.4.1
urllib3==1.25.8
Werkzeug==1.0.0
yarl==1.4.2
